fix(controller): use smart pointer for mapping loader
